requirements.txt
flask
flask_mail
flask_jwt_extended
flask_bcrypt
flask_sqlalchemy
python-dotenv
gunicorn
flask-cors
email
smtplib
ssl